Fixed overflow caused by vendors with > 255 items. Number of items actually sent to a client of particular class is limited to 150 by the client.

+++ src/server/game/Entities/Creature/Creature.h
@@ -338,7 +338,7 @@ struct VendorItemData
         return m_items[slot];
     }
     bool Empty() const { return m_items.empty(); }
-    uint8 GetItemCount() const { return m_items.size(); }
+    uint32 GetItemCount() const { return m_items.size(); }
     void AddItem(uint32 item, int32 maxcount, uint32 ptime, uint32 ExtendedCost)
     {
         m_items.push_back(new VendorItem(item, maxcount, ptime, ExtendedCost));
